requirements.txt
alembic==0.9.10
apache-airflow==1.10.3
Babel==2.7.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.5.3
croniter==0.3.30
defusedxml==0.6.0
dill==0.2.9
docutils==0.14
Flask==1.0
Flask-Admin==1.5.3
Flask-AppBuilder==1.12.3
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
json-merge-patch==0.2
lockfile==0.12.2
lxml==4.3.4
Mako==1.0.13
Markdown==2.6.11
MarkupSafe==1.1.1
numpy==1.16.4
ordereddict==1.1
pandas==0.24.2
pendulum==1.4.4
psutil==5.6.3
Pygments==2.4.2
pymongo==3.8.0
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.2
PyYAML==5.1.1
requests==2.22.0
setproctitle==1.1.10
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.19
tabulate==0.8.3
tenacity==4.12.0
text-unidecode==1.2
thrift==0.11.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.4.0
